Move --isolate file into the isolate root.
Isolate archival happens relative to the directory of the "--isolate"
file. When the pathing in the "isolate" command was correctly
updated, the "--isolate" file location was not, causing Kitchen builds
to continue to fail.

This moves the "--isolate" file into the same directory tree as the
actual content being isolated.

BUG=chromiuim:722121
TEST=expectations
R=sergeyberezin@chromium.org, smut@chromium.org

Change-Id: I3b4523eb265e66f14f0e36dbc003be14c1496dbd
Reviewed-on: https://chromium-review.googlesource.com/506523
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: smut <smut@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>