chromite: Explicitly fetch depot_tools and put in path.
cbuildbot requires depot_tools in the path for a variety of reasons,
but this isn't the norm for swarming builds. So, fetch it and stick it
into the path before invoking cbuildbot.

Also adjust the code for fetching chromite so that we do the same
thing for both. Put both depot_tools and chromite in the cache,
instead of storing them under 'start_dir'.

BUG=chromium:720171

Change-Id: Ib1fbbc8859bb83a4593dcdaab02175079315bbfb
Reviewed-on: https://chromium-review.googlesource.com/506784
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>