Turn off swarming for the linux layout tests temporarily.
It appears that when we run the layout tests under swarming,
the script that merges the JSON results deletes the current
working directory, which is the main writable directory used
by remote_run. This is causing failures and other weirdness.
For now, I'm turning off swarming, so that we can then properly
debug things and figure out what the right thing to do is.

R=iannucci@chromium.org
TBR=tansell@chromium.org, mcgreevy@chromium.org
BUG=721466

Change-Id: Ib6ee007cf162335116b564214e2a0ee901860593
Reviewed-on: https://chromium-review.googlesource.com/508158
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Michael McGreevy <mcgreevy@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Tim 'mithro' Ansell <tansell@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>