Handle OSError raised by api.git in findit module.
This may happen if the cwd given to the function does not exist.
Whether the cwd is correct or not is a different issue.

Since the purpose of the code being amended is to record data for
computing optimizations and not the test itself, failing at these points
should not abort the recipe run.

R=iannucci,chanli,lijeffrey,stgao

Bug:
Change-Id: Ib9f2ff3a0de7d44326ed65f9388051f1722d5b7d
Reviewed-on: https://chromium-review.googlesource.com/528535
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>