isolate: Add support for percentage experiments for exparchive.
In https://crbug.com/689735 the perfbots where converted whole sale to
use exparchive instead of batcharchive. This significantly improve
performance.

However, we are worried that exparchive might not be faster in *all*
cases, so we want to run an experiment with a percentage of runs on a
builder using exparchive rather than batcharchive.

BUG=689735,692940,598990

Change-Id: Id00f35567edcdc83c77a38c8d0f86e3da1096519
Reviewed-on: https://chromium-review.googlesource.com/526874
Commit-Queue: Tim 'mithro' Ansell <tansell@chromium.org>
Reviewed-by: Michael McGreevy <mcgreevy@chromium.org>