[remote_run] Do emergency cleanup on full disk.
The "cleanup_temp" stage of "remote_run" will raise an Exception if the
disk is full. This prevents any of "remote_run"'s cleanup logic from
executing, since cleanup happens well after this point.

Allow the cleanup function to be executed with a lighter context, and
call this in response to a full disk error. This should automatically
repair some existing systems that have full disks, but wouldn't if a
basic cleanup was able to happen.

BUG=chromium:725631
TEST=local
R=iannucci@chromium.org, mmoss@chromium.org

Change-Id: I6ab17cc0648e681ae104960f444dd1d96b089004
Reviewed-on: https://chromium-review.googlesource.com/530075
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>