linux_chromium_rel_ng: Enable exparchive for 5% of builds.
In https://crbug.com/689735 the perfbots where converted whole sale to
use exparchive instead of batcharchive. This significantly improve
performance.

However, we are worried that exparchive might not be faster in *all*
cases, so we want to run an experiment with a percentage of runs on a
builder using exparchive rather than batcharchive.

BUG=689735,692940,598990

Change-Id: I2cbead4bbcb0c4f68b1ba111284b1521ff5d72fd
Reviewed-on: https://chromium-review.googlesource.com/532767
Commit-Queue: Tim 'mithro' Ansell <tansell@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Paweł Hajdan Jr. <phajdan.jr@chromium.org>