Re-land "Implement builder_defaults, mixins, and inline bots for builders.pyl.""
This reverts commit c453ad81 and re-lands 30df994d with the needed fix.
Specifically, a very new internal master was using the 'bot_data' and I
had thought no one was using that. This CL handles that case, and also
adds a better diagnostic message so that if something fails we get the
path to the builders.pyl file.

TBR=nodir@chromium.org

Change-Id: I309f6bf9400d18622034a8a9f81cdec34e1dda22
Reviewed-on: https://chromium-review.googlesource.com/545165
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>