Update wasm.llvm build source to llvm.googlesource.com
The buildbot syncs code from llvm.googlesource.com, but the build was
triggered by commits to the github mirror. This could lead to the two mirrors
being out of sync, which would cause build failures.

This change makes the build source the same as the trigger source.

Bug:
Change-Id: I053921a218cbbb845c56560a16f0338364b0df17
Reviewed-on: https://chromium-review.googlesource.com/556381
Reviewed-by: Brad Nelson <bradnelson@chromium.org>
Commit-Queue: Brad Nelson <bradnelson@chromium.org>