Shuffle bots around in the tryserver.chromium.mac iOS pools.
As a part of managing the capacity crunch on iOS for
crbug.com/739556, this CL moves 8 bots from the
ios-simulator-eg pool and 2 bots from the ios-device-xcode-clang
pool to the ios-device pool, doubling its capacity from 10
devices to 20. Hopefully that will be enough to handle the
(temporary?) increase in load. With crbug.com/739565
we should be able to replace the two builders and add five more
to ios-device-xcode-clang and possibly move five back from
ios-device to ios-device-xcode-clang, leaving each at 15 bots
and hopefully that'll be enough (otherwise we can move things
around again).

R=smut@google.com, sergeyberezin@chromium.org, pschidt@chromium.org
BUG=739556, 739565

Change-Id: I3499d352f3a34e3a7a895db869db84f4bc10008b
Reviewed-on: https://chromium-review.googlesource.com/560671
Reviewed-by: smut <smut@google.com>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>