Use new path to dart-sdk for flutter recipes
The dart-flutter integration build (see [0]) started to suddently fail
some time ago.  It turned out that the failure started to occur after an
unrealted issue which caused a the buildbot slave to do a fresh
checkout.  After the fresh checkout "flutter/travis/analyze.sh" failed
due to missing "dartanalyzer".

The reason turned out to be that "gclient runhooks" is lazily
downloading the dart-sdk to a *different* location now.  Older buildbot
slave checkouts might still have the old dart-sdk in the checkout (and
therefore continue to work, just using a stale dart-sdk) but new checkouts
will fail.

This CL changes the flutter recipes to set PATH pointing to the new
dart-sdk location (which was changed in [1]).

[0] https://uberchromegw.corp.google.com/i/client.dart.internal/builders/flutter-engine-linux
[1] https://github.com/flutter/buildroot/commit/214524470c62170604261d1bed7270015c54a9b4

Bug:
Change-Id: Ie99531cdcda06a263286e62dcc37a109e695d272
Reviewed-on: https://chromium-review.googlesource.com/565503
Reviewed-by: Jason Simmons <jsimmons@chromium.org>
Commit-Queue: Martin Kustermann <kustermann@google.com>