Modify the way analyze is implemented in the chromium recipes.
Previously, if a file matches something in the trybot exlusion
list, we wouldn't call `mb analyze` at all, and would exit early,
assuming everything needed to rebuild. However, this would let
bugs in mb or gn get through the CQ accidentally (particularly
leading to bugs in analyze that would show up on subsequent tryjobs).

This CL changes the behavior so that we *always* call `mb analyze`.
If analyze errors out, then we raise an error. However, if it runs
properly, and a file matches the exclusion rules, we then ignore
the results from MB.

One effect of this change is that a build will now have a new
"analyze_matched_exclusion" step as well as the "analyze" step,
if the exclusions were matched.

R=phajdan.jr@chromium.org
BUG=736941

Change-Id: Ic2c3d2f72f94e32e128605c94e7e3cced0181c4f
Reviewed-on: https://chromium-review.googlesource.com/566087
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Paweł Hajdan Jr. <phajdan.jr@chromium.org>