Force builders that don't use the hermetic toolchain to choose an SDK.
Builders that don't use the hermetic toolchain cannot use the default Chromium
SDK, since that is dependent on the hermetic toolchain. There are two builders
that don't use the hermetic toolchain: webkit testers and webrtc. The former
expects the 10.10 SDK and the latter expects the 10.11 SDK.

Bug:chromium:669240
Change-Id: I23db631294cffa559d06fc3b104a577e2713869d
Reviewed-on: https://chromium-review.googlesource.com/565656
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>