Configure a new master.tryserver.chromium.chromiumos.
This CL defines a new master that will take some of the load off of
master.tryserver.chromium.linux. Specifically, this will eventually be
the home of all of the non-memory, non-variable, non-fyi chromeos bots
current on tryserver.chromium.linux:

- chromeos_amd64-generic
- chromeos_amd64-generic_chromium_compile_only (a CQ bot)
- chromeos_daisy
- chromeos_daisy_chromium_compile_only (a CQ bot)
- linux_chromium_chromeos_compile_dbg (should be a CQ bot, but currently isn't)
- linux_chromium_chromeos_dbg
- linux_chromium_chromiumos_rel (a CQ bot)

(Note that the _ng suffixes are gone).

R=nodir@chromium.org
BUG=741164

Change-Id: I939e8ea166725831e456510a3802f92e9e39ba07
Reviewed-on: https://chromium-review.googlesource.com/567553
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>