Rework buildbot-tool to better be able to keep things in sync.
Apparently people have been modifying the files generated by
buildbot-tool instead of modifying builders.pyl and
re-running buildbot-tool to keep them in sync.

This change reworks buildbot-tool to add a `check` command to
make sure the files are in sync, and adds a (for now commented-out)
presubmit check to enforce this.

Once I land a follow-up CL to update the files and a matching one
for the internal repos, I can uncomment the presubmit checks.

R=dnj@chromium.org
BUG=732748

Change-Id: I3763d4c73da8a489bb520d3299cd54813a8e4f52
Reviewed-on: https://chromium-review.googlesource.com/573140
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Daniel Jacques <dnj@chromium.org>