Pass through upload_test_results parameter for swarming
SwarmingIsolatedScriptTest has an optional parameter
upload_test_results. By passing this param from the swarming config, we
allow Telemetry (and other users of generate_isolated_script) to
delineate which test results should be uploaded to the server. We
default this to True if not defined since that is the current default
in SwarmingIsolatedScriptTest.__init__.

BUG=chromium:748638

Change-Id: Ia8fa52a3d9bf2f72a65d28ae383953775305b579
Reviewed-on: https://chromium-review.googlesource.com/586972
Commit-Queue: Ashley Enstad <ashleymarie@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>