Remove win x86 builders, we'll only build on x64 for now.
Since we don't want to support building directly on 32-bit
bots, the 32-bit builders on crashpad are effectively useless.
This CL removes them.

This means that the 32-bit-build-on-32-bit-hosts configuration
is (still) untested and will be until we are running the tests
under swarming.

TBR=mark@chromium.org
BUG=743139

Change-Id: I7faebadf4e6f9b349e057f2a7840cd8477e1e2b8
Reviewed-on: https://chromium-review.googlesource.com/592535
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>