Update crashpad recipe to supporting swarming.
In order to move the bots onto LUCI-based tryservers
and test on non-x64 platforms, we need to be running
the tests under swarming rather than running them locally
on the build machines.

This CL adds in the initial support for this; the tests
are run as before, but if the file //build/swarming_tests_spec.pyl
exists, the tests listed in it will be run under swarming as well.

BUG=697636, 743139

Change-Id: Ia5c6550a99f3e0c5138874547d8d86f68e180141
Reviewed-on: https://chromium-review.googlesource.com/580375
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>