Reland "V8: Use builder cache in V8 recipe" (part 1)
This introduces a checkout_root field in the v8 module to be used
as parent folder of the checkouts.

In a follow up, this'll switch to builder_cache where applicable.

This is a reland of 03ed79d87940024b36e7093938d2dfa456a8c18a
Original change's description:
> V8: Use builder cache in V8 recipe
> 
> Bug: 747960
> Change-Id: I3e5f3e77d8a7e049b871dd1f7b6af6e659eede1b
> Reviewed-on: https://chromium-review.googlesource.com/598007
> Reviewed-by: Daniel Jacques <dnj@chromium.org>
> Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
> Commit-Queue: Michael Achenbach <machenbach@chromium.org>

Bug: 747960
Change-Id: Ieda2b912e39211f31eca9637f09d6de7683117c8
Reviewed-on: https://chromium-review.googlesource.com/598909
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Paweł Hajdan Jr. <phajdan.jr@chromium.org>