WebRTC: Fix bug causing Debug builders to trigger testers
In https://chromium-review.googlesource.com/616736 a bug was introduced
that caused the debug builders to take over the triggering of the testers
instead of the Release builders.

Using platform+build config as key in the dictionary will restore the previous
behavior.

This would have been caught if we had the expectations around, but
it has been decided they're undesired, so we lost that protection.

BUG=756741
TBR=iannucci@chromium.org, ehmaldonado@chromium.org

Change-Id: I3054542929b7cbc3b32ee31ef72fdea81c7b3981
Reviewed-on: https://chromium-review.googlesource.com/620588
Reviewed-by: Henrik Kjellander <kjellander@chromium.org>
Commit-Queue: Henrik Kjellander <kjellander@chromium.org>