Increase the period for wpt-importer jobs to 10 minutes
Even though the longer jobs can take more than 3 hours, it's OK to have
a relatively short period because I've discovered that pending jobs are
cleared when new jobs start.

So this means that we'll normally expect wpt-importer to retry once
every 10 minutes when there's nothing to do, and when there is something
to do, 30-40 pending builds might build up but will be cleared at the
start of a new build.

Bug: 755756
Change-Id: Ia0319753f75c44a89fb8436da523e3543142e47b
Reviewed-on: https://chromium-review.googlesource.com/630137
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>