Add trigger script to swarming recipe module
This script can be used to change how swarming tasks get triggered.
The recipe code will call the script with the same arguments that get
normally passed to `swarming.py trigger`, with a few modifications. It
will treat the return value and corresponding resulting output json
files as if they came from swarming.py trigger.

This directly enables https://chromium-review.googlesource.com/c/chromium/src/+/629962,
which I'll be working to land after this CL lands.
It triggers multiple shards of the same isolate, telemetry_perf_tests,
each with `id: buildXX` dimensions, which isn't possible using the current infrastructure.
In the future we'll be changing which device we trigger at runtime depending on which
devices are healthy when the script is running.



Change-Id: I3343d0d89fb34b2439c0641a120cfae069b2b2c0
Reviewed-on: https://chromium-review.googlesource.com/629437
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>