Add better stats on gtest swarming collection tasks.
This will indicate that the pending time is the time for the
longest-pending shard, and also indicates the maximum shard time of all
the shards that were collected.

If there was more than one shard, each time measurement says which shard
it belongs to, as well.

R=charliea@chromium.org, jbudorick@chromium.org, vadimsh@chromium.org

Bug: 758339
Change-Id: Ifbf4f1b30ac7825a9bdeba317946582b1b1b391f
Reviewed-on: https://chromium-review.googlesource.com/642491
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>