Add jumbo builders to chromium.fyi
The "jumbo" build is a GN configuration where we compile
multiple C++ (or Obj-C++) files in a single translation unit;
this should lead to significant speedups in compile times.

However, we actually want to run benchmarks to validate this.
The "Jumbo {Linux,Mac,Win}" bots on the chromium.fyi waterfall
will build the same things as the {"Linux x64", "Mac", "Win x64"}
bots on the chromium waterfall, which will hopefully give
us something like an apples-to-apples comparison of compile times.

R=jbudorick@chromium.org, brucedawson@chromium.org, bratell@opera.com
BUG=745862

Change-Id: If750fdcc67e87283211a66ec19add21047af398f
Reviewed-on: https://chromium-review.googlesource.com/647320
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>