[cipd_bootstrap_v2] Canary new Python bundle.
This is a reland of a14395c9ac122b181c3121fbad37b1bd0261eabd.

This new version of Python is far more static than chromium9. It also
turns out that the reason for revert was because of a bug in the unit
test, not the Python bundle. See:

https://chromium.googlesource.com/chromium/tools/build/+/980c9fa32dd36501286458e419cb9ede895330d3

BUG=chromium:741001
TEST=dev
R=iannucci@chromium.org

Change-Id: Ibe67efb6402338daf3b07d50bbe46a2c972f8d7f
Reviewed-on: https://chromium-review.googlesource.com/653255
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>