Histogram Pipeline - Move add_reserved_diagnostics call to upload_perf_dashboard_results.py.
Previously, this was setup and called out to in the chromium_tests/steps.py, but this has
been moved inside upload_perf_dashboard_results.py just before the upload step to the
dashboard. This simplifies the recipe side of things, and means when this script migrates src
side we get unit testing coverage, etc.

Change-Id: I31540f2d776fa9dcaf43926c3b9d7d365c5a5857
Reviewed-on: https://chromium-review.googlesource.com/685946
Commit-Queue: Simon Hatch <simonhatch@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Emily Hanley <eyaich@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>