Only generate the dashboard link for chartjson.
Currently, results_dashboard.SendResults tries to generate a link to a
chromeperf dashboard report, but that assumes that the data is chartjson.
We aren't sure how useful this link is, and handling HistogramSet json here
would require importing the HistogramSet library from catapult, so this CL
disables this link for HistogramSet results.

Change-Id: I8ed42a1bad87aedf51a16d402ced47e27c39835f
Reviewed-on: https://chromium-review.googlesource.com/706110
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>