Temporarily disable 3 of the 4 copies of the layout test results uploads.
When we archive the results of the layout tests, we do it 4 times,
to provide the "latest results" as a zip file and as a browsable
tree, and to provide the results for the build number at a stable
location as a zip file and as a browsable tree.

This CL temporarily disables all but the stable zip file URL, in
order to reduce the upload load.

R=hinoka@chromium.org, jchinlee@chromium.org
BUG=776334

Change-Id: I95775d74a398c2e3d28040d4bdcb567baeeadf7e
Reviewed-on: https://chromium-review.googlesource.com/729154
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Katie Thomas <katthomas@google.com>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>