Revert back to the original (4-copy) version of the archive script.
For now, this CL restores the original version, so we can be sure
everything is working the way it was. I would still like to replace
this with smarter versions ASAP that at least did a cloud2cloud
copy for two of the copies.

TBR=hinoka@chromium.org, jchinlee@chromium.org
BUG=776334

Change-Id: Ie5f27b98e69078d1376e30c28a1f66e1fb13f8e2
Reviewed-on: https://chromium-review.googlesource.com/729316
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>