Rework chromium.chromiumos waterfalls and builders.
As part of cleaning up how we test CrOS upstream on the
chromium waterfalls (see the bug), this adds new builders
to the chromium.chromiumos and tryserver.chromium.chromiumos
waterfalls.

As part of this, this CL switches builders to a new naming
scheme that is based (loosely) on the iOS naming scheme:

target_os[-{descriptor}]*[-compile]-(dbg|rel)[-arch]

and uses those same names on the continuous waterfall and the
tryserver waterfall. For example:

  linux-chromeos-rel 
  linux-chromeos-compile-dbg
  linux-chromeos-dbg
  chromeos-amd64-generic-rel
  chromeos-betty-rel
  chromeos-daisy-rel

With "linux-chromeos-rel", this indicates that this is a linux
build in the "chromeos" configuration, as opposed to a build
that is meant to run on an actual CrOS image, e.g. a simplechrome
or ebuild build. (To make things even more confusing, the
"chromeos" config sets the GN arg target_os="chromeos", but I
don't think we should replicate that mistake here). 

With "chromeos-amd64-generic-rel", "amd64-generic" is the descriptor,
which corresponds to the "amd64-generic" board. We'll use the
board to imply the architecture, which is why "chromeos-daisy-rel"
isn't "chromeos-daisy-rel-arm".  

R=friedman@chromium.org, smut@google.com
BUG=732531, 777020

Change-Id: I08ef3e0d31399f4cd93c3e48c22868d15b420d10
Reviewed-on: https://chromium-review.googlesource.com/731832
Reviewed-by: smut <smut@google.com>
Reviewed-by: Elliott Friedman <friedman@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>