Tweak archive_layout_test_results script again.
Apparently GSUtilCopyFile() appends the basename of the src
file to the destination (`gs_base`) unless you also specify
`dest_filename` as an arg. Since we want the basename to be
the same for the uploaded files, we just need to not append
the basename to the `gs_base` arg.

Also, remove the _unittest.py file, it no longer tests anything
interesting, and check the return codes from gsutil.

TBR=hinoka@chromium.org
BUG=776334

Change-Id: I8886e4e06826182f122bba25c03af267d9b95c40
Reviewed-on: https://chromium-review.googlesource.com/731862
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>