Revert "Move downstream recipe trybots to production"
This reverts commit 1f49cf4a2bd63f0dc207ccd120d078cb32758125.

Reason for revert: 

The downstream bots create a deadlock: when the upstream change (correctly) requires a non-trivial roll, the bot will fail. The only way for it to succeed is to roll the downstream with the upstream change, but it can't be done since the upstream CL hasn't landed yet.

Case in point: https://crrev.com/c/738848

Original change's description:
> Move downstream recipe trybots to production
> 
> Bug: 644609
> Change-Id: I90e60e5260d136b5a38425929f5592acc70593ea
> Reviewed-on: https://chromium-review.googlesource.com/702315
> Reviewed-by: Aaron Gable <agable@chromium.org>
> Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
> Commit-Queue: Paweł Hajdan Jr. <phajdan.jr@chromium.org>

TBR=iannucci@chromium.org,agable@chromium.org,phajdan.jr@chromium.org

# Not skipping CQ checks because original CL landed > 1 day ago.

Bug: 644609
Change-Id: I4d90d7dd4235c7ff29af5f8e5d384fdda386ff35
Reviewed-on: https://chromium-review.googlesource.com/739803
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>
Commit-Queue: Sergey Berezin <sergeyberezin@chromium.org>