Make WebRTC bots upload wav files from failed AQ tests.
This will make it a lot easier to work out what's going on
on the bots when audio quality tests fail. We will be able
to listen to recordings by clicking a link in the failed
job and unzipping the gathered wav files.

This change makes the test receive a new flag
--webrtc_save_audio_recordings_in=/some/tmpdir. The temp
dir will disappear from the bot after the build is
complete.

https://chromium-review.googlesource.com/c/chromium/src/+/728242
implements the flag and should land before this CL.

Bug:767857
Change-Id: I93bc2b358fd75c509fe125a83616a33c2f9c6bbd
Reviewed-on: https://chromium-review.googlesource.com/725292
Commit-Queue: Patrik Höglund <phoglund@chromium.org>
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>