Simplifies Kythe corpus to allow for merging
The corpus labels are ignored by the previous Xrefs pipeline
but used by the new Kythe pipeline. The labels need to be the
same in order for the current merging scheme to work.

To enable multi-config support in the future, revert to use
these labels again.

Bug:665742
Change-Id: I32d1588359afa8ebcef32780ddc7419a17b12e6c
Reviewed-on: https://chromium-review.googlesource.com/746661
Commit-Queue: Emma Söderberg <emso@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>