Stop running BlinkTest() on Mac, remove force_mac_toolchain configs.
The chromium.webkit Mac builders are now all running the tests
under swarming, so we can stop running the tests locally.

As part of this, this stops running the BlinkTest() step
on mac chromium.webkit bots. That step is what explicitly 
runs the layout tests locally on the bot; the remote/swarmed 
invocation is already being triggered by the src-side 
//testing/buildbot/*.json files.

This also removes the "Chromium Mac 10.11" and
"Chromium Mac 10.11 Force Mac Toolchain" bots from the 
chromium.fyi waterfall, and drops the 
'force_mac10.10_toolchain' toolchain config, which is
no longer needed.

This also re-arranges the tryserver.blink builders so that
they are now sharing the 10.12 pool. This pool only has
3 builders, which may result in pending builds at times, but
we can reimage some of the 10.10/10.11 bots and add them to
the pool if need be.

With all of this, we can now assume that everything is 
building on 10.12 with the 10.12 SDK.

R=qyearsley@chromium.org, tansell@chromium.org, erikchen@chromium.org
BUG=524758

Change-Id: I34892c35430f2a88c7553a38a2ee3e7f1e7b7e2b
Reviewed-on: https://chromium-review.googlesource.com/749623
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>