Add compile-only targets for ANGLE try bots.
The intent for this change is to allow ANGLE to stop testing the x64
versions of the tests on the ANGLE CQ. The x64 full testers will be
replaced on the CQ with compile-only versions, and the full testers
will remain on the FYI bots.

This also renames the debug testers to make it more clear they are
compile-only. No new FYI bot mirrors are needed because the new
compile-only configurations are merely mirrors of the FYI builders.

The old x64 targets are left around. This is necessary to prevent
breakage, and also we could keep them around as non-default try jobs
for tests.

Bug: 786044
Tbr: tandrii@chromium.org
Change-Id: I9c0aa9f6582461a7654144f95943d6bf6a0def9d
Reviewed-on: https://chromium-review.googlesource.com/775147
Commit-Queue: Jamie Madill <jmadill@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>