Add a flag in recipe to allow tests run on passed-in MacOS version.
We have some tests failing on macOS-10.13,  which is not ready for corp
machines yet. Developers cannot fix the test until that happens.  Add a
workaround in iOS recipe to allow tests run on passed-in MacOS version
machines for iOS simulator tests.

The workaround will allow tests to pass in a macOS version flag in the
configuration file for simulators:
{
  "app": "chrome_test",
  "host os": "Mac-10.12",
  "device type": "iPhone 7",
  "os": "11.0"
}
Bug:785185
Change-Id: I475e35e12f90ff4bb1fe7cd5d8f40dabf3de87c9
Reviewed-on: https://chromium-review.googlesource.com/775361
Commit-Queue: Menglu Huang <huangml@chromium.org>
Reviewed-by: smut <smut@google.com>