Merge chromium.clang builder and tester bots
These bots were configured as separate builders and testers, but now
that we have swarming, the build packaging and uploading steps create
unnecessary overhead. Consolidating these bots will also reduce the
number of boxes to check in the buildbot web UI.

The only remaining builder-only bot after this change is "CFI Linux CF".
It's not clear if this was intentionally configured to not run tests,
but in any case, this patch doesn't touch it.

These are the tester VMs that are no longer used:
  vm12-m1
  vm30-m1
  vm190-m1
  vm310-m1
  vm311-m1
  vm312-m1
  vm313-m1
  vm961-m1
  vm962-m1
  vm970-m1
  vm978-m1

R=dpranke@chromium.org, hans@chromium.org

Bug: 765420, 587879
Change-Id: Ice1dd50194c70fb57d1d2eae5f0a36d7dd3accd6
Reviewed-on: https://chromium-review.googlesource.com/782619
Reviewed-by: Hans Wennborg <hans@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Reid Kleckner <rnk@chromium.org>