[swarming] Fix usage of api.properties
This code assumes that api.properties only contains 'simple' values like (str,
int, bool), but that's not the case. In the event that a property contains
a JSON Object, it will return a FrozenDict, which cannot be serialized to JSON.
`thaw()` is built to return a mutable copy of api.properties' contents.

R=agable@chromium.org, vadimsh@chromium.org

Bug: 791127
Change-Id: Ia3c466fc940b9d3ee2b66db82bbac45412505e7c
Reviewed-on: https://chromium-review.googlesource.com/804934
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>