Clean up handling of test args and --gtest_filter, esp. on retries.
The interaction between --gtest_filter and --test-launcher-filter-file
is somewhat complicated, and the recipe-side code was making it even
more complicated. In particular, in cases where a test was configured
to use a filter file, the recipe would throw away the filter file
argument on a retry, and that would cause problems when we changed
the filter files as part of a patch.

This CL rationalizes the changes a bit, so that we always use
--gtest-filter for the failures to retry; eventually, we should
change this so that we always use --test-launcher-filter-file so
that we don't have to worry about the size of the command line, but
that requires more work to ensure that the file is plumbed through
to swarming properly.

In addition, this removes an argument that is apparently no longer
needed in the chromium_android API, making the control flow slightly
simpler (one fewer place to override things), and removes the use
of a static method from the GTestTest class; really the whole class
needs to be removed, but that's for a different CL.

R=jbudorick@chromium.org
BUG=667004, 791211

Change-Id: I1d5899cbbbb2eeee72aba058f87d73bb8b02cb76
Reviewed-on: https://chromium-review.googlesource.com/804682
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>