[luci] Fix depot_tools location on $PATH on LUCI.
depot_tools is pretty badly organized; it has tools which are
meant to be on $PATH for developers (like vpython), but also
has critical tools like gclient and git cl.

In buildbot there's an unpinned copy of depot_tools on $PATH.
However, we want the recipes-pinned copy of depot_tools ahead
of it on $PATH. Instead of doing something sane, we added a
hack to the recipe engine to make tools like vpython always
at the front of $PATH, regardless of what 'env_prefixes' is
set to in recipes.

LUCI no longer has this hack, but it ALSO doesn't have an
unpinned copy of depot_tools on $PATH. This means that it's
safe (and saneish) to put depot_tools at the end of $PATH.

This logic is encapsulated in the "depot_tools.on_path"
context manager. This CL doesn't change any expectations
because nothing seems to set the `is_luci` bit to true
in the test cases, but that also indicates that this
CL has no effect for recipes running on buildbot.

R=tandrii@chromium.org

Bug:
Change-Id: I23305dd54afbbce2e3ebd00980319ff812659494
Reviewed-on: https://chromium-review.googlesource.com/827641
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>