Add custom step name to layout test archive uploads.
Currently it's only possible to upload one set of layout test
results per build. This is a problem if you want to run the
tests twice, with different configurations. This CL changes things
so that the step name will be included in the path to the upload
if we're using a step name other than 'webkit_tests' or
'webkit_layout_tests' (ignoring the with/without patch part).

R=jbudorick@chromium.org, qyearsley@chromium.org
BUG=786554

Change-Id: I198add586d8fedc9cd6f7ecb6572b6e9b3030476
Reviewed-on: https://chromium-review.googlesource.com/833092
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>