Rely on bot configurations to trigger the layout tests.
This change switches the chromium_trybot recipe to rely
completely on the src-side bot configs to determine whether
to run the layout tests on a change or not. As part of this,
we delete all of the custom logic, and we change the
win7_chromium_rel_ng bot to mirror the 'WebKit Win7'
builder as well as the other builders, and change the
mac_chromium_rel_ng bot to mirror the 'WebKit Mac10.12'
builder; this'll have the net effect of causing the
layout tests to run on Win7 and 10.12 when analyze thinks
it needs to do so.

We still need the custom BlinkTest() step for the
WebKit Android (Nexus4) builder, which isn't yet
configured to run under swarming :(.

R=jbudorick@chromium.org, qyearsley@chromium.org
BUG=524758

Change-Id: Ib29341954a70c070e2500425c15c857e691e0eb4
Reviewed-on: https://chromium-review.googlesource.com/833527
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>