Don't submit CLs in auto_roll_webrtc_deps if the revision diff is <100
This adds the flag from https://webrtc-review.googlesource.com/36320

Currently there are around 250 Chromium CLs per day on average, so
this would mean ~2.5 roll CLs a day (previously 24), and less busy
times are automatically adjusted for.

--cq-over=100 means that "Roll chromium_revision (524047:524105)"
would not be submitted because the revision difference is below 100.
The CLs will still be uploaded and tested, as a rudimentary version
of an FYI waterfall of DEPS rolls.

Bug: webrtc:8688
Change-Id: Ia1b248d76252aa53c7443a3d80e517454d206d02
Reviewed-on: https://chromium-review.googlesource.com/843977
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Oleh Prypin <oprypin@chromium.org>