Remove steps.GTestTest.
GTestTest in its current incarnation exists almost exclusively as a factory
and a wrapper class for SwarmingGTestTest and LocalGTestTest. In those roles,
though, it provides little value -- clients know whether they're creating
a swarming test or not based on the arguments they provide to the constructor,
and the wrapping doesn't modify any existing functionality. This CL removes
the class given that limited value & its tendency to confuse and/or frustrate
readers or modifiers of steps.py.

Change-Id: I5b180cd503ae95b6dc4bdfded80d944dfdb205c5
Bug:
Reviewed-on: https://chromium-review.googlesource.com/842091
Commit-Queue: John Budorick <jbudorick@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>