Simplifies crashpad_win, and moves build to GN
Moves the Windows build to be GN instead of gyp. I suspect something
won't work as I can't get the recipe to run locally, but hey, who knows.

There's only one slave in the pool for Windows builds at the moment,
vm254-m3. As a result, there's no point in separating the wow64 from the
main x64, as we need to generate and build x64 and x86 on the wow64 one
anyway. As total run time is dominated by buildbot script setup and the
compilation, just do x64 and x86 wow64 on the same machine. If we ever
figure out how to run on swarming to get x86 on x86, then we can still
do that from this bot in the x86 build dir.

Bug: crashpad:79
Change-Id: Ie8438778074004689c992cc59749ef31e45b8092
Reviewed-on: https://chromium-review.googlesource.com/905516
Commit-Queue: Scott Graham <scottmg@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>