Fix typo in archive layout test script
Python doesn't interpret (x) as a tuple; it ignores the parantheses.
This CL changes it to a list, so that it's checking if a string is in
a list, rather than a string in a string.

The script should still work as is, but is technically incorrect as is.

Bug: 808167
Change-Id: Ie9ad5c0a0ea49094b59bee821be3551a3849ba0a
Reviewed-on: https://chromium-review.googlesource.com/907696
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>