V8: Retire gyp handling when running MB
Last usage of passing the gyp config file to MB has been removed on the
V8 side.

This is also compatible with V8's beta and stable builders. They're all
on GN in the current beta/stable versions tested. The parameter is not
used in that case.

Bug: 772804
Change-Id: Idec5a918988525232aae7a0349d7c55e269baa3b
Reviewed-on: https://chromium-review.googlesource.com/912831
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>