WebRTC: Trigger a simple task in Android Perf (swarming)
Trigger a simple task in Android Perf (swarming) instead of a GTest task.

As it is, it complains because we don't produce an output.json file
with the execution time and success status for each test we ran, and makes
the build purple even though the task succeeded.

Triggering a simple task, we don't check for that file and simply collect
the task results.

Bug: 755660
Change-Id: I11e2332815d022e3a8dafe7651d0e16fa10cda37
Reviewed-on: https://chromium-review.googlesource.com/913578
Reviewed-by: Patrik Höglund <phoglund@chromium.org>
Commit-Queue: Edward Lesmes <ehmaldonado@chromium.org>