Add presubmit test which verifies trybots' configuration.
The chromium_trybot recipe does these checks at runtime, but in order
to prevent accidents it's crucial that they be done at presubmit time.

It's not feasible to write this as a unit test either inside the
recipes or by importing files like trybots.py from outside the recipe
infrastructure. However, masters_recipes_test already contains all of
the needed infrastructure; it uses a preexisting contract with the
chromium and chromium_trybot recipes to dump all of the waterfalls'
and trybots' configurations as JSON.

A new test has been added to masters_recipes_test asserting:

  1) The waterfall bots, which the trybots mirror, must all exist.

  2) Keys (like "chromium_apply_config") which are asserted to be
  identical among all of a trybot's mirrored builders when the recipe
  runs, are now checked to be identical at presubmit time.

Comments have been added to make it clear that a bit of logic in this
test must be identical to logic inside the recipe.

Five trybots were found to be misconfigured while adding this new test
and their specifications have been removed. They will be removed from
chromium/src as well. Some cleanups were made to the GPU waterfalls to
keep the test's logic simple.

The new test has been verified to catch errors accidentally introduced
recently, and which motivated writing the test in the first place.

Bug: 814087
Change-Id: I49589a095b433ba751dfd7752be08cb8c3cc815f
Reviewed-on: https://chromium-review.googlesource.com/927868
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>